requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2==2.9.9
redis==5.0.1
aiogram==3.2.0
aiohttp==3.9.1
python-dotenv==1.0.0
jinja2==3.1.2
python-multipart==0.0.6